Format consistency fix in UTCTime conversion
When converting UTCTime to JSON, subseconds are omitted if subseconds
are exactly 0 (no fractional part). This results in a format
"Y-m-dTH-m-sZ" which is in contrast with the "Y-m-dTH-m-s.SSSZ" format
that is normally created.

This patch adds the ".000" part when there is no fractional part in the
UTCTime that is to be converted.